Switch to LWRP-driven cookbook
Use attributes only for the global HAProxy configuration file.  All
other configuration will be done with the new LWRP, using the
multi-config-file functionality in the latest version of
`Consul-HAProxy`.  This allows for multiple cookbooks to reside on the
same node, all specifiying `consul-haproxy` configs that will be
composed together at runtime.